refactor(backend): extract file operations to FileRepository, backend.ts now coordinator only
