+++ src/backend/backend.ts
@@ -2,8 +2,6 @@ import {
   AllusionDB_SQL,
   deserializeBoolean,
   deserializeDate,
-  EpValues,
-  Files,
   LocationNodes,
   Locations,
   LocationTags,
@@ -13,29 +11,13 @@ import {
   ExtraProperties as DbExtraProperties,
   SavedSearches,
   SearchCriteria,
-  FileTags,
   SearchGroups,
 } from './schemaTypes';
 import SQLite from 'better-sqlite3';
 
-import {
-  Kysely,
-  sql,
-  SelectQueryBuilder,
-  SqlBool,
-  AnyColumn,
-  Insertable,
-  Expression,
-} from 'kysely';
+import { Kysely, sql, SelectQueryBuilder, AnyColumn, Insertable, Expression } from 'kysely';
 import { migrateToLatest } from './config';
-import {
-  initDB,
-  generateSeed,
-  getSqliteMaxVariables,
-  computeBatchSize,
-  PadString,
-  stableHash,
-} from './db';
+import { initDB, generateSeed, getSqliteMaxVariables, computeBatchSize } from './db';
 import { DataStorage } from 'src/api/data-storage';
 import {
   OrderBy,
@@ -45,22 +27,23 @@ import {
   Cursor,
   IndexableType,
 } from 'src/api/data-storage-search';
-import { ExtraProperties, ExtraPropertyDTO } from 'src/api/extraProperty';
+import { ExtraPropertyDTO } from 'src/api/extraProperty';
 import { FileDTO, FileStats } from 'src/api/file';
 import { FileSearchDTO, SearchGroupDTO } from 'src/api/file-search';
-import { generateId, ID } from 'src/api/id';
+import { ID } from 'src/api/id';
 import { LocationDTO, SubLocationDTO } from 'src/api/location';
 import { ROOT_TAG_ID, TagDTO } from 'src/api/tag';
 import { jsonArrayFrom } from 'kysely/helpers/sqlite';
 import { IS_DEV } from 'common/process';
 import { UpdateObject } from 'kysely/dist/cjs/parser/update-set-parser';
 import { SemanticSearchOptions, SemanticSearchStatus } from 'src/api/semantic-search';
-import { applyFileFilters, applyPagination, PaginationOptions } from './query-builder';
+import { PaginationOptions } from './query-builder';
 import { SemanticRepository } from './repositories/SemanticRepository';
 import { TagRepository } from './repositories/TagRepository';
 import { LocationRepository } from './repositories/LocationRepository';
 import { SearchRepository } from './repositories/SearchRepository';
 import { ExtraPropertyRepository } from './repositories/ExtraPropertyRepository';
+import { FileRepository } from './repositories/FileRepository';
 
 // Use to debug perfomance.
 const USE_TIMING_PROXY = IS_DEV;
@@ -72,13 +55,12 @@ export default class Backend implements DataStorage {
   #dbPath!: string;
   #notifyChange!: () => void;
   #restoreEmpty!: () => Promise<void>;
-  // Seed used to have deterministic order when order by random
-  #seed: number = generateSeed();
   #semantic!: SemanticRepository;
   #tags!: TagRepository;
   #locations!: LocationRepository;
   #searches!: SearchRepository;
   #extraProperties!: ExtraPropertyRepository;
+  #files!: FileRepository;
 
   constructor() {
     // Must call init() before using to init the properties.
@@ -107,19 +89,27 @@ export default class Backend implements DataStorage {
       await migrateToLatest(db, { jsonToImport });
     }
 
+    this.#tags = new TagRepository(this.#db, this.MAX_VARS, this.#notifyChange);
     this.#semantic = new SemanticRepository(
       db,
       sqlite,
       (ids) => this.fetchFilesByID(ids),
-      (criteria, pagOptions) => this.queryFiles(criteria, pagOptions),
+      (criteria, pagOptions) => this.#files.queryFiles(criteria, pagOptions),
     );
-    this.#tags = new TagRepository(this.#db, this.MAX_VARS, this.#notifyChange);
     this.#locations = new LocationRepository(this.#db, this.MAX_VARS, this.#notifyChange);
     this.#searches = new SearchRepository(this.#db, this.MAX_VARS, this.#notifyChange);
     this.#extraProperties = new ExtraPropertyRepository(
       this.#db,
       this.MAX_VARS,
       this.#notifyChange,
+    );
+    this.#files = new FileRepository(
+      this.#db,
+      this.#tags,
+      this.#semantic,
+      this.MAX_VARS,
+      this.#notifyChange,
+      generateSeed(),
     );
 
     if (mode === 'migrate' || mode === 'readonly') {
@@ -160,7 +150,7 @@ export default class Backend implements DataStorage {
   }
 
   async setSeed(seed?: number): Promise<void> {
-    this.#seed = seed ?? generateSeed();
+    return this.#files.setSeed(seed);
   }
 
   async fetchTags(): Promise<TagDTO[]> { return this.#tags.fetchTags(); }
@@ -168,41 +158,11 @@ export default class Backend implements DataStorage {
   async preAggregateJSON(): Promise<void> { return this.#tags.preAggregateJSON(); }
 
   async queryFiles<Q extends SelectQueryBuilder<any, any, any>>(
-    criteria: ConditionGroupDTO<FileDTO> = { conjunction: 'and', children: [] },
+    criteria: ConditionGroupDTO<FileDTO>,
     pagOptions: PaginationOptions,
     modifyQuery?: (qb: Q) => Q,
   ): Promise<FileDTO[]> {
-    pagOptions.seed = this.#seed;
-    if (this.#tags.isQueryDirty) {
-      await this.preAggregateJSON();
-    }
-    const dbWithTemp = this.#db.withTables<{
-      fileTagAggregatesTemp: {
-        fileId: ID;
-        tags: ID[];
-      };
-      fileEpAggregatesTemp: {
-        fileId: ID;
-        extraProperties: EpValues[];
-      };
-    }>();
-    // Apply the filter criterias expressions to the files QueryBuilder and execute the query.
-    let query;
-    query = dbWithTemp
-      .selectFrom('files')
-      .leftJoin('fileTagAggregatesTemp as ft', 'ft.fileId', 'files.id')
-      .leftJoin('fileEpAggregatesTemp as fe', 'fe.fileId', 'files.id')
-      .selectAll('files')
-      .select(['ft.tags', 'fe.extraProperties']);
-    query = applyFileFilters(query, criteria);
-    query = await applyPagination(this.#db, query, pagOptions);
-    if (modifyQuery) {
-      query = modifyQuery(query as any);
-    }
-
-    const files = (await query.execute()).map(mapToDTO);
-    const shouldReverse = pagOptions.pagination === 'before' && pagOptions.cursor !== undefined;
-    return shouldReverse ? files.reverse() : files;
+    return this.#files.queryFiles(criteria, pagOptions, modifyQuery);
   }
 
   async fetchFiles(
@@ -214,16 +174,15 @@ export default class Backend implements DataStorage {
     cursor?: Cursor,
     extraPropertyID?: ID,
   ): Promise<FileDTO[]> {
-    console.info('SQLite: Fetching all files...', cursor);
-    return this.queryFiles(undefined, {
+    return this.#files.fetchFiles(
       order,
-      direction: fileOrder,
+      fileOrder,
       useNaturalOrdering,
       limit,
       pagination,
       cursor,
       extraPropertyID,
-    });
+    );
   }
 
   async searchFiles(
@@ -236,16 +195,16 @@ export default class Backend implements DataStorage {
     cursor?: Cursor,
     extraPropertyID?: ID,
   ): Promise<FileDTO[]> {
-    console.info('SQLite: Searching files...', cursor, criteria);
-    return this.queryFiles(criteria, {
+    return this.#files.searchFiles(
+      criteria,
       order,
-      direction: fileOrder,
+      fileOrder,
       useNaturalOrdering,
       limit,
       pagination,
       cursor,
       extraPropertyID,
-    });
+    );
   }
 
   async semanticSearchByText(query: string, options?: SemanticSearchOptions): Promise<FileDTO[]> {
@@ -273,24 +232,11 @@ export default class Backend implements DataStorage {
   }
 
   async fetchFilesByID(ids: ID[]): Promise<FileDTO[]> {
-    console.info('SQLite: Fetching files by ID...', ids);
-    return this.queryFiles(undefined, { order: 'dateAdded' }, (query) =>
-      query.where('id', 'in', ids),
-    );
+    return this.#files.fetchFilesByID(ids);
   }
 
   async fetchFilesByKey(key: keyof FileDTO, values: IndexableType): Promise<FileDTO[]> {
-    console.info('SQLite: Fetching files by key...');
-    if (!['tags', 'extraProperties', 'extraPropertyIDs'].includes(key)) {
-      if (!Array.isArray(values)) {
-        values = [values as string | number | Date];
-      }
-      return this.queryFiles(undefined, { order: 'dateAdded' }, (query) =>
-        query.where(key, 'in', values),
-      );
-    }
-    console.error('fetchFilesByKey error: Key or values not supported.');
-    return [];
+    return this.#files.fetchFilesByKey(key, values);
   }
 
   async fetchLocations(): Promise<LocationDTO[]> {
@@ -471,35 +417,8 @@ export default class Backend implements DataStorage {
     return this.#tags.createTag(tag);
   }
 
-  // Creates many files at once, and checks for duplicates in the path they are in
   async createFilesFromPath(path: string, filesDTO: FileDTO[]): Promise<void> {
-    console.info('SQLite: Creating files...', path, filesDTO.length);
-
-    if (filesDTO.length === 0) {
-      return;
-    }
-    const { files } = normalizeFiles(filesDTO);
-    const FILES_BATCH_SIZE = computeBatchSize(this.MAX_VARS, files[0]);
-    await this.#db.transaction().execute(async (trx) => {
-      for (let i = 0; i < files.length; i += FILES_BATCH_SIZE) {
-        const batch = files.slice(i, i + FILES_BATCH_SIZE);
-        try {
-          await trx
-            .insertInto('files')
-            .values(batch)
-            .onConflict((oc) => oc.doNothing())
-            .execute();
-        } catch (error) {
-          console.error(`Failed to insert files batch at index ${i}:`, error);
-        }
-      }
-    });
-
-    this.#semantic.enqueueSemanticEmbeddings(filesDTO.map((file) => file.id));
-
-    this.#tags.isQueryDirty = true;
-    this.#notifyChange();
-    console.info('SQLite: Files created successfully');
+    return this.#files.createFilesFromPath(path, filesDTO);
   }
 
   async createLocation(location: LocationDTO): Promise<void> {
@@ -526,108 +445,7 @@ export default class Backend implements DataStorage {
   }
 
   async saveFiles(filesDTO: FileDTO[]): Promise<void> {
-    console.info('SQLite: Saving files...', filesDTO);
-    if (filesDTO.length === 0) {
-      return;
-    }
-
-    const { fileIds, files, fileTags, epVal } = normalizeFiles(filesDTO);
-
-    // Compute batch sizes. To use the maximum number of vars SQLite can handle per query.
-    const DELETE_BATCH_SIZE = this.MAX_VARS;
-    const FILES_BATCH_SIZE = computeBatchSize(this.MAX_VARS, files[0]);
-    const FILE_TAGS_BATCH_SIZE = computeBatchSize(this.MAX_VARS, fileTags[0]);
-    const EP_VALUES_BATCH_SIZE = computeBatchSize(this.MAX_VARS, epVal[0]);
-
-    await this.#db.transaction().execute(async (trx) => {
-      // Create unique temp table names.
-      const tempSuffix = generateId();
-      const tempFiles = `files_temp_${tempSuffix}`;
-      const tempFileTags = `file_tags_temp_${tempSuffix}`;
-      const tempEpValues = `ep_values_temp_${tempSuffix}`;
-
-      try {
-        // Create temp tables form a copy of the actual tables.
-        await sql`CREATE TEMP TABLE ${sql.id(tempFiles)} AS SELECT * FROM files WHERE 0`.execute(
-          trx,
-        );
-        await sql`CREATE TEMP TABLE ${sql.id(
-          tempFileTags,
-        )} AS SELECT * FROM file_tags WHERE 0`.execute(trx);
-        await sql`CREATE TEMP TABLE ${sql.id(
-          tempEpValues,
-        )} AS SELECT * FROM ep_values WHERE 0`.execute(trx);
-        // Insert files into temp files table
-        for (let i = 0; i < files.length; i += FILES_BATCH_SIZE) {
-          const batch = files.slice(i, i + FILES_BATCH_SIZE);
-          await trx
-            .insertInto(tempFiles as any)
-            .values(batch)
-            .execute();
-        }
-        // Delete previous fileTags and epValues, it is quicker to delete all from related files and insert them in bulk.
-        if (fileIds.length > 0) {
-          for (let i = 0; i < fileIds.length; i += DELETE_BATCH_SIZE) {
-            const batchIds = fileIds.slice(i, i + DELETE_BATCH_SIZE);
-            await trx.deleteFrom('fileTags').where('fileId', 'in', batchIds).execute();
-            await trx.deleteFrom('epValues').where('fileId', 'in', batchIds).execute();
-          }
-        }
-        // Insert fileTags into temp table
-        if (fileTags.length > 0) {
-          for (let i = 0; i < fileTags.length; i += FILE_TAGS_BATCH_SIZE) {
-            const batch = fileTags.slice(i, i + FILE_TAGS_BATCH_SIZE);
-            await trx
-              .insertInto(tempFileTags as any)
-              .values(batch)
-              .execute();
-          }
-        }
-        // Insert epValues into temp table
-        if (epVal.length > 0) {
-          for (let i = 0; i < epVal.length; i += EP_VALUES_BATCH_SIZE) {
-            const batch = epVal.slice(i, i + EP_VALUES_BATCH_SIZE);
-            await trx
-              .insertInto(tempEpValues as any)
-              .values(batch)
-              .execute();
-          }
-        }
-        // Transfer from temp tables
-        // Upsert FILES
-        upsertTable(
-          this.MAX_VARS,
-          trx,
-          'files',
-          sql`SELECT * FROM ${sql.id(tempFiles)} WHERE true`,
-          ['id'],
-          ['dateAdded'],
-          files[0],
-        );
-        // Insert FileTags
-        if (fileTags.length > 0) {
-          await sql`
-          INSERT INTO file_tags 
-          SELECT * FROM ${sql.id(tempFileTags)}
-        `.execute(trx);
-        }
-        // Insert EpValues
-        if (epVal.length > 0) {
-          await sql`
-          INSERT INTO ep_values 
-          SELECT * FROM ${sql.id(tempEpValues)}
-        `.execute(trx);
-        }
-        this.#tags.isQueryDirty = true;
-        console.info('SQLite: Files saved successfully');
-      } finally {
-        // Clean temp table.
-        await sql`DROP TABLE IF EXISTS ${sql.id(tempFiles)}`.execute(trx);
-        await sql`DROP TABLE IF EXISTS ${sql.id(tempFileTags)}`.execute(trx);
-        await sql`DROP TABLE IF EXISTS ${sql.id(tempEpValues)}`.execute(trx);
-      }
-    });
-    this.#notifyChange();
+    return this.#files.saveFiles(filesDTO);
   }
 
   async saveLocation(location: LocationDTO): Promise<void> {
@@ -710,10 +528,7 @@ export default class Backend implements DataStorage {
   }
 
   async removeFiles(files: ID[]): Promise<void> {
-    console.info('SQLite: Removing files...', files);
-    // Cascade delte in other tables deleting from files table.
-    await this.#db.deleteFrom('files').where('id', 'in', files).execute();
-    this.#notifyChange();
+    return this.#files.removeFiles(files);
   }
 
   async removeLocation(location: ID): Promise<void> {
@@ -740,307 +555,39 @@ export default class Backend implements DataStorage {
   }
 
   async addTagsToFiles(tagIds: ID[], criteria?: ConditionGroupDTO<FileDTO>): Promise<void> {
-    console.info('SQLite: Add tags to filtered files...', criteria, tagIds);
-    // Subquery tipado correctamente
-    let fileSubquery = this.#db.selectFrom('files').select('files.id as fileId');
-    fileSubquery = applyFileFilters(fileSubquery, criteria);
-
-    // Crear valores de tags como CTE o subquery
-    await this.#db
-      .insertInto('fileTags')
-      .columns(['fileId', 'tagId'])
-      .expression(() => {
-        // Usar raw SQL para el cross join con los valores
-        const tagValues = tagIds.map((id) => `SELECT '${id}' as tag_id`).join(' UNION ALL ');
-
-        return this.#db
-          .selectFrom(fileSubquery.as('matchedFiles'))
-          .crossJoin(sql`(${sql.raw(tagValues)})`.as('tagValues'))
-          .select(['matchedFiles.fileId', sql<number>`tag_values.tag_id`.as('tagId')])
-          .where(sql<SqlBool>`true`);
-      })
-      .onConflict((oc) => oc.doNothing())
-      .execute();
-
-    this.#tags.isQueryDirty = true;
+    return this.#files.addTagsToFiles(tagIds, criteria);
   }
 
   async removeTagsFromFiles(tagIds: ID[], criteria?: ConditionGroupDTO<FileDTO>): Promise<void> {
-    console.info('SQLite: Remove tags from filtered files...', criteria, tagIds);
-
-    let fileSubquery = this.#db.selectFrom('files').select('files.id');
-    fileSubquery = applyFileFilters(fileSubquery, criteria);
-
-    await this.#db
-      .deleteFrom('fileTags')
-      .where('fileId', 'in', fileSubquery)
-      .where('tagId', 'in', tagIds)
-      .execute();
-
-    this.#tags.isQueryDirty = true;
+    return this.#files.removeTagsFromFiles(tagIds, criteria);
   }
 
   async clearTagsFromFiles(criteria?: ConditionGroupDTO<FileDTO>): Promise<void> {
-    let fileSubquery = this.#db.selectFrom('files').select('files.id');
-    fileSubquery = applyFileFilters(fileSubquery, criteria);
-
-    await this.#db.deleteFrom('fileTags').where('fileId', 'in', fileSubquery).execute();
-
-    this.#tags.isQueryDirty = true;
+    return this.#files.clearTagsFromFiles(criteria);
   }
 
   async countFiles(
     options?: { files?: boolean; untagged?: boolean },
     criteria?: ConditionGroupDTO<FileDTO>,
   ): Promise<[fileCount: number | undefined, untaggedFileCount: number | undefined]> {
-    console.info('SQLite: Counting files...', options, criteria);
-    const result: [number | undefined, number | undefined] = [undefined, undefined];
-    if (options?.files) {
-      let totalQuery = this.#db
-        .selectFrom('files')
-        .select(({ fn }) => fn.count<number>('files.id').as('count'));
-      totalQuery = criteria ? applyFileFilters(totalQuery, criteria) : totalQuery;
-      const totalResult = await totalQuery.executeTakeFirst();
-      result[0] = totalResult?.count ?? 0;
-    }
-
-    if (options?.untagged) {
-      let untaggedQuery = this.#db
-        .selectFrom('files')
-        .leftJoin('fileTags as ft', 'ft.fileId', 'files.id')
-        .where('ft.fileId', 'is', null)
-        .select(({ fn }) => fn.count<number>('files.id').as('count'));
-      untaggedQuery = criteria ? applyFileFilters(untaggedQuery, criteria) : untaggedQuery;
-      const untaggedResult = await untaggedQuery.executeTakeFirst();
-      result[1] = untaggedResult?.count ?? 0;
-    }
-    return result;
+    return this.#files.countFiles(options, criteria);
   }
 
-  /** Compare the given disk files with the database files for the given location. */
   async compareFiles(
     locationId: ID,
     diskFiles: FileStats[],
   ): Promise<{ createdStats: FileStats[]; missingFiles: FileDTO[] }> {
-    const dbWithTemp = this.#db.withTables<{
-      tempDiskFiles: Omit<FileStats, 'dateModified' | 'dateCreated'> & {
-        dateModified: number;
-        dateCreated: number;
-      };
-    }>();
-    // first insert all missing files into a temp table for easier and db optimized querying
-    // use unique table name for concurrency
-    const tempSuffix = generateId();
-    const tempDiskFilesName = `temp_disk_files_${tempSuffix}`;
-    const tempDiskFiles = sql
-      .table(tempDiskFilesName)
-      .as('tempDiskFiles') as unknown as 'tempDiskFiles';
-
-    await sql`
-      CREATE TEMP TABLE ${sql.id(tempDiskFilesName)} (
-        absolute_path   TEXT PRIMARY KEY,
-        ino            TEXT NOT NULL,
-        size           INTEGER NOT NULL,
-        date_modified   INTEGER NOT NULL,
-        date_created    INTEGER NOT NULL
-      ) WITHOUT ROWID;
-    `.execute(this.#db);
-
-    const DISK_FILES_BATCH_SIZE = computeBatchSize(this.MAX_VARS, diskFiles[0]);
-    await dbWithTemp.transaction().execute(async (trx) => {
-      for (let i = 0; i < diskFiles.length; i += DISK_FILES_BATCH_SIZE) {
-        const batch = [];
-        const end = Math.min(i + DISK_FILES_BATCH_SIZE, diskFiles.length);
-        for (let j = i; j < end; j++) {
-          const f = diskFiles[j];
-          batch.push({
-            absolutePath: f.absolutePath,
-            ino: f.ino,
-            size: f.size,
-            dateModified: serializeDate(f.dateModified),
-            dateCreated: serializeDate(f.dateCreated),
-          });
-        }
-        await trx
-          .insertInto(tempDiskFilesName as 'tempDiskFiles')
-          .values(batch)
-          .execute();
-      }
-    });
-
-    // find created files, (the ones present in disk but not in db)
-    const createdStats: FileStats[] = (
-      await dbWithTemp
-        .selectFrom(tempDiskFiles)
-        .leftJoin('files', (join) =>
-          join
-            .onRef('files.absolutePath', '=', 'tempDiskFiles.absolutePath')
-            .on('files.locationId', '=', locationId),
-        )
-        .where('files.id', 'is', null)
-        .selectAll('tempDiskFiles')
-        .execute()
-    ).map((df) => ({
-      absolutePath: df.absolutePath,
-      ino: df.ino,
-      size: df.size,
-      dateModified: deserializeDate(df.dateModified),
-      dateCreated: deserializeDate(df.dateCreated),
-    }));
-
-    // find missing files, (the ones present in db but not in disk)
-    const missingFiles = await this.queryFiles(
-      undefined,
-      { order: 'id' },
-      (query: SelectQueryBuilder<AllusionDB_SQL & { tempDiskFiles: FileStats }, 'files', any>) => {
-        return query
-          .leftJoin(tempDiskFiles, (join) =>
-            join.onRef('tempDiskFiles.absolutePath', '=', 'files.absolutePath'),
-          )
-          .where('files.locationId', '=', locationId)
-          .where('tempDiskFiles.absolutePath', 'is', null);
-      },
-    );
-    // clean temp table
-    await sql`DROP TABLE IF EXISTS ${sql.id(tempDiskFilesName)}`.execute(this.#db);
-
-    return { createdStats, missingFiles };
+    return this.#files.compareFiles(locationId, diskFiles);
   }
 
-  /** Find possible matches in the database for the given missing files based on their metadata. */
   async findMissingDBMatches(
     missingFiles: FileDTO[],
   ): Promise<Array<[missingFileId: ID, dbMatch: FileDTO]>> {
-    if (missingFiles.length === 0) {
-      return [];
-    }
-
-    const dbWithTemp = this.#db.withTables<{
-      tempMissingFiles: {
-        id: string;
-        name: string;
-        ino: string;
-        width: number | null;
-        height: number | null;
-        dateCreated: number;
-      };
-      fileTagAggregatesTemp: {
-        fileId: ID;
-        tags: ID[];
-      };
-      fileEpAggregatesTemp: {
-        fileId: ID;
-        extraProperties: EpValues[];
-      };
-    }>();
-
-    // first insert all missing files into a temp table for easier and db optimized querying
-    // use unique table name for concurrency
-    const tempMissingName = `temp_missing_files_${generateId()}`;
-    const tempMissingFiles = sql
-      .table(tempMissingName)
-      .as('tempMissingFiles') as unknown as 'tempMissingFiles';
-
-    await sql`
-      CREATE TEMP TABLE ${sql.id(tempMissingName)} (
-        id TEXT PRIMARY KEY,
-        name TEXT,
-        ino TEXT,
-        width INTEGER,
-        height INTEGER,
-        date_created INTEGER
-      ) WITHOUT ROWID;
-    `.execute(this.#db);
-
-    const BATCH_SIZE = computeBatchSize(this.MAX_VARS, missingFiles[0]);
-    await dbWithTemp.transaction().execute(async (trx) => {
-      for (let i = 0; i < missingFiles.length; i += BATCH_SIZE) {
-        const batch = [];
-        const end = Math.min(i + BATCH_SIZE, missingFiles.length);
-        for (let j = i; j < end; j++) {
-          const f = missingFiles[j];
-          batch.push({
-            id: f.id,
-            name: f.name,
-            ino: f.ino,
-            width: f.width,
-            height: f.height,
-            dateCreated: serializeDate(f.dateCreated),
-          });
-        }
-        await trx
-          .insertInto(tempMissingName as 'tempMissingFiles')
-          .values(batch)
-          .execute();
-      }
-    });
-
-    // Compare metadata of two files to determine whether the files are (likely to be) identical
-    // same logic as areFilesIdenticalBesidesName but in DB for optimization to trasverse all files.
-    const matches = await dbWithTemp
-      .selectFrom(tempMissingFiles)
-      .innerJoin('files', (join) =>
-        join
-          .onRef('files.id', '!=', 'tempMissingFiles.id')
-          .on((eb) =>
-            eb.or([
-              eb('files.ino', '=', eb.ref('tempMissingFiles.ino')),
-              eb.and([
-                eb('files.width', '=', eb.ref('tempMissingFiles.width')),
-                eb('files.height', '=', eb.ref('tempMissingFiles.height')),
-                eb('files.dateCreated', '=', eb.ref('tempMissingFiles.dateCreated')),
-              ]),
-            ]),
-          ),
-      )
-      .leftJoin('fileTagAggregatesTemp as ft', 'ft.fileId', 'files.id')
-      .leftJoin('fileEpAggregatesTemp as fe', 'fe.fileId', 'files.id')
-      .selectAll('files')
-      .select(['ft.tags', 'fe.extraProperties', 'tempMissingFiles.id as missingSourceId'])
-      // prioritize matches by name first, then by id to have a stable order
-      .orderBy('tempMissingFiles.id')
-      .orderBy(sql`CASE WHEN files.name = ${sql.ref('tempMissingFiles.name')} THEN 0 ELSE 1 END`)
-      .execute();
-
-    // clean temp table
-    await sql`DROP TABLE IF EXISTS ${sql.id(tempMissingName)}`.execute(this.#db);
-
-    // multiple matches can be found for the same missing file, keep the best one (first by name)
-    const uniqueMatches = new Map<ID, FileDTO>();
-    for (const row of matches) {
-      if (!uniqueMatches.has(row.missingSourceId as ID)) {
-        const { missingSourceId, ...fileData } = row;
-        uniqueMatches.set(missingSourceId as ID, mapToDTO(fileData));
-      }
-    }
-
-    // return entries for compatibility with worker mode.
-    return Array.from(uniqueMatches.entries()).map(([missingId, matchedFile]) => [
-      missingId,
-      matchedFile,
-    ]);
+    return this.#files.findMissingDBMatches(missingFiles);
   }
 
   async clear(): Promise<void> {
-    console.info('SQLite: Clearing database...');
-    /*
-    const tables = await this.#db
-      .selectFrom('sqlite_master' as any)
-      .select('name')
-      .where('type', '=', 'table')
-      .where('name', 'not like', 'sqlite_%')
-      .execute();
-
-    for (const { name } of tables) {
-      if (name === 'kysely_migration' || name === 'kysely_migration_lock') {
-        continue;
-      }
-      await this.#db.deleteFrom(name as any).execute();
-    } */
-
-    // Empy the tables with a large database takes too long, instead create an emprty DB,
-    // reinit and restore it at startup relying in the backup-scheduler checkAndRestoreDB behaviour.
-    await this.#restoreEmpty();
+    return this.#files.clear(this.#restoreEmpty);
   }
 }
 
@@ -1065,39 +612,6 @@ function createTimingProxy(obj: Backend): Backend {
       return original;
     },
   });
-}
-
-function mapToDTO(dbFile: FileDTO | { [x: string]: any }): FileDTO {
-  // convert data into FileDTO format
-  const extraPropertyIDs: ID[] = [];
-  const extraProperties: ExtraProperties = {};
-  for (const ep of dbFile.extraProperties ?? []) {
-    extraPropertyIDs.push(ep.epId);
-    const val = ep.textValue ?? ep.numberValue; // ?? ep.timestampValue;
-    if (val !== null) {
-      extraProperties[ep.epId] = val;
-    }
-  }
-  return {
-    id: dbFile.id,
-    ino: dbFile.ino,
-    locationId: dbFile.locationId,
-    relativePath: dbFile.relativePath,
-    absolutePath: dbFile.absolutePath,
-    tagSorting: dbFile.tagSorting,
-    dateAdded: deserializeDate(dbFile.dateAdded),
-    dateModified: deserializeDate(dbFile.dateModified),
-    dateModifiedOS: deserializeDate(dbFile.dateModifiedOS),
-    dateLastIndexed: deserializeDate(dbFile.dateLastIndexed),
-    dateCreated: deserializeDate(dbFile.dateCreated),
-    name: dbFile.name,
-    extension: dbFile.extension,
-    size: dbFile.size,
-    width: dbFile.width,
-    height: dbFile.height,
-    tags: dbFile.tags ?? [],
-    extraProperties: extraProperties,
-  };
 }
 
 ///////////////////
@@ -1279,61 +793,4 @@ function normalizeSavedSearches(sourceSearches: FileSearchDTO[]) {
     searchGroups,
     searchCriteria,
   };
-}
-
-function normalizeFiles(sourceFiles: FileDTO[]) {
-  const fileIds: ID[] = [];
-  const files: Insertable<Files>[] = [];
-  const fileTags: Insertable<FileTags>[] = [];
-  const epVal: Insertable<EpValues>[] = [];
-
-  for (const file of sourceFiles) {
-    const fileId = file.id;
-    fileIds.push(fileId);
-    files.push({
-      id: fileId,
-      ino: file.ino,
-      locationId: file.locationId,
-      relativePath: file.relativePath,
-      absolutePath: file.absolutePath,
-      tagSorting: file.tagSorting,
-      name: file.name,
-      extension: file.extension,
-      size: file.size,
-      width: file.width,
-      height: file.height,
-      dateAdded: serializeDate(file.dateAdded),
-      dateModified: serializeDate(file.dateModified),
-      dateModifiedOS: serializeDate(file.dateModifiedOS),
-      dateLastIndexed: serializeDate(file.dateLastIndexed),
-      dateCreated: serializeDate(file.dateCreated),
-    });
-    // file_tags (tags relations)
-    for (const tagId of Array.isArray(file.tags) ? file.tags : []) {
-      fileTags.push({
-        fileId: fileId,
-        tagId: tagId,
-      });
-    }
-    // ep_values  (extra properties relations)
-    for (const [epId, value] of Object.entries(file.extraProperties)) {
-      // TODO: Maybe should fetch the ExtraProperties types to assign the type based on
-      // the extra property definition, but since the DTO types do not overlap for now, this
-      // is good enough.
-      if (typeof value === 'number') {
-        epVal.push({
-          fileId,
-          epId,
-          numberValue: value,
-        });
-      } else {
-        epVal.push({
-          fileId,
-          epId,
-          textValue: value,
-        });
-      }
-    }
-  }
-  return { fileIds, files, fileTags, epVal };
 }
